requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httptools==0.6.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pyasn1==0.5.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
